+++ packages/components/src/tools-panel/test/index.js
@@ -316,7 +316,7 @@ describe( 'ToolsPanel', () => {
 			expect( announcement ).toHaveAttribute( 'aria-live', 'assertive' );
 		} );
 
-		it( 'should render optional panel item when corresponding value is updated externally', async () => {
+		it( 'should render optional panel item when value is updated externally and panel has an ID', async () => {
 			const ToolsPanelOptional = ( { toolsPanelItemValue } ) => {
 				const itemProps = {
 					attributes: { value: toolsPanelItemValue },
@@ -325,7 +325,7 @@ describe( 'ToolsPanel', () => {
 					onDeselect: jest.fn(),
 					onSelect: jest.fn(),
 				};
-				altControlProps.attributes.value = toolsPanelItemValue;
+
 				return (
 					<ToolsPanel { ...defaultProps }>
 						<ToolsPanelItem { ...itemProps }>
@@ -342,7 +342,41 @@ describe( 'ToolsPanel', () => {
 
 			rerender( <ToolsPanelOptional toolsPanelItemValue={ 100 } /> );
 
-			const controlRerendered = screen.queryByText( 'Optional control' );
+			const controlRerendered = screen.getByText( 'Optional control' );
+
+			expect( controlRerendered ).toBeInTheDocument();
+		} );
+
+		it( 'should render optional item when value is updated externally and panelId is null', async () => {
+			// This test partially covers: https://github.com/WordPress/gutenberg/issues/47368
+			const ToolsPanelOptional = ( { toolsPanelItemValue } ) => {
+				const itemProps = {
+					attributes: { value: toolsPanelItemValue },
+					hasValue: () => !! toolsPanelItemValue,
+					label: 'Alt',
+					onDeselect: jest.fn(),
+					onSelect: jest.fn(),
+				};
+
+				// The null panelId below simulates the panel prop when there
+				// are multiple blocks selected.
+				return (
+					<ToolsPanel { ...defaultProps } panelId={ null }>
+						<ToolsPanelItem { ...itemProps }>
+							<div>Optional control</div>
+						</ToolsPanelItem>
+					</ToolsPanel>
+				);
+			};
+
+			const { rerender } = render( <ToolsPanelOptional /> );
+			const control = screen.queryByText( 'Optional control' );
+
+			expect( control ).not.toBeInTheDocument();
+
+			rerender( <ToolsPanelOptional toolsPanelItemValue={ 99 } /> );
+
+			const controlRerendered = screen.getByText( 'Optional control' );
 
 			expect( controlRerendered ).toBeInTheDocument();
 		} );